requirements.txt
google-cloud-datastore==1.11.0
future==0.18.2
coverage==5.0.3
